+++ src/Vankosoft/ApplicationInstalatorBundle/Command/ClearInstallCommand.php
@@ -27,7 +27,7 @@ EOT
     protected function execute( InputInterface $input, OutputInterface $output ): int
     {
         $applicationName    = $input->getArgument( 'application' );
-        $appSetup           = $this->getContainer()->get( 'vs_application.installer.setup_application' );
+        $appSetup           = $this->get( 'vs_application.installer.setup_application' );
         
         $outputStyle        = new SymfonyStyle( $input, $output );